# contracts/openapi/orders.yaml
openapi: 3.1.0
info:
  title: Servvia orders API
  version: 0.1.0-as-built
  description: |
    Contract of the EXISTING NestJS implementation (apps/api) for order
    creation, listing, status transitions and the kiosk Stripe helpers.
    Written so a replacement service can match the wire behaviour exactly,
    including known defects (`x-known-issues`). Every statement cites source
    via `x-source` (repo-relative path:line). Items not provable from code
    carry `x-unverified: true`.

    ## Global behaviour

    * Prefix `/api` (apps/api/src/main.ts:151).
    * Global `ValidationPipe({ whitelist: true, transform: true })`
      (apps/api/src/main.ts:150): unknown body properties are silently
      STRIPPED (not rejected; `forbidNonWhitelisted` is off), including
      inside nested item/modifier objects. Validation failure -> 400
      `{ statusCode: 400, message: string[], error: "Bad Request" }`.
    * No global exception filter: errors use the NestJS default
      `{ statusCode, message, error }`, EXCEPT exceptions constructed with an
      object body (STALE_PRICE), whose body is that object verbatim with no
      `statusCode` key (@nestjs/common 11 `HttpException.createBody`).
      Unhandled errors -> 500 `{ statusCode: 500, message: "Internal server error" }`.
    * Request pipeline order (NestJS): middleware (SecurityHeaders, CSRF) ->
      guards -> pipes (validation) -> handler. So CSRF 403 precedes
...
    * All responses carry the security headers from
...

...

...

...

...
servers:
...
paths:
  /kiosk/orders:
    post:
      tags: [kiosk]
      operationId: createKioskOrder
      summary: Public, prepaid kiosk order (Stripe Terminal)
      description: |
        Unauthenticated. The order is persisted `status = confirmed`,
        `source = kiosk`, `serviceMode = dine_in`, `guests = null`,
        `confirmedAt = now`, `paymentProviderTransactionId =
        stripePaymentIntentId`. Modifiers use the LEGACY name-based
        resolution (see `components.x-modifier-resolution.nonStrict`).

        **Exact processing order** (orders.service.ts:102-238):
        1. Venue by id (`findUnique`, isActive NOT checked) -> 404 `Venue not found`.
        2. Look up existing order by `(venueId, idempotencyKey)`.
        3. If none: table pre-check `validateTableForOrder(venueId, tableId, tableNumber)`
           (see `components.x-table-resolution`).
        4. Price every line (`resolveOrderItems`, non-strict) — runs for
           replays too.
        5. Upsert kiosk system actor (audit only).
        6. If an existing order was found -> idempotent outcome (see
           `components.x-idempotency`); returns or throws here.
        7. `checkExpectedPrices` (STALE_PRICE 409).
        8. Payment-reference pre-check (`rejectIfPaymentReferenceReused`) -> 409.
        9. Table re-resolution (same rules as 3).
        10. `computeTotals` (422 on unsupported venue tax profile).
        11. `verifyKioskPayment`: retrieves the PaymentIntent from Stripe
            (skipped only when `NODE_ENV === 'test'`):
            missing/`sk_test_dummy` key -> 503 `Stripe payment is not configured`;
            retrieve error -> 400 `Unable to verify the Stripe payment`;
            `status !== 'succeeded'` -> 400 `Stripe payment has not succeeded`;
            `amount !== server totalCents` OR currency not `nzd`
            (case-insensitive) -> 409 `Stripe payment amount or currency does
            not match the order`.
        12. Persist (transaction: table `FOR UPDATE` recheck, order, items,
            legacy POS outbox row, printer jobs, KDS delivery row). P2002
            conflicts are recovered (see x-idempotency).
        13. Audit `CREATE_ORDER`, WebSocket broadcast, return 201.

        **Production (Story 12.5):** this route is off whenever `NODE_ENV`
        is anything other than exactly `development` or `test` (including
        unset). A request that gets past CSRF gets a plain `404 Not Found`,
        as for a route that does not exist, before rate limiting, body
        validation and any service, database or Stripe work. There is no
        setting that re-enables it in production.
      x-source: apps/api/src/orders/orders.controller.ts:63-68
      x-implementation: apps/api/src/orders/orders.service.ts:102-238
      x-rate-limit:
        limit: 30
        windowSeconds: 60
        x-source: apps/api/src/orders/orders.controller.ts:64-65
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: |
            New order, or the ORIGINAL order on a matching idempotent replay
            (same 201 status; there is no header or field distinguishing a
            replay). Shapes differ: new -> `OrderCreated` (items +
            posSyncRecord, no table); replay -> `OrderReplay` (items only).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OrderCreated'
                  - $ref: '#/components/schemas/OrderReplay'
        '400':
          description: |
            DTO validation, or business rule: `Invalid or inactive table`;
            `MenuItem with ID <id> not found`; `An order must contain at
            least one item`; `Each selected modifier must supply a name`;
            `"<name>" is not a valid option for this item`; `Unable to verify
            the Stripe payment`; `Stripe payment has not succeeded`;
            `stripePaymentIntentId is required for payment verification`
            (unreachable: DTO already enforces non-empty).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '404':
          description: |
            Development and test: `Venue not found` (venue id unknown,
            orders.service.ts:1038-1044).
            Production (Story 12.5): the route does not exist. Returned for
            every request that gets past CSRF with a well-formed JSON body
            (or none), empty or invalid, before rate limiting and body
            validation, so never 429 or 503 and never a validation 400. A
            malformed or oversized body is refused (400/413) by the shared
            body parser before routing, as for any path. The body carries no
            route, provider or configuration detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
              examples:
                venueNotFound:
                  value: { statusCode: 404, message: Venue not found, error: Not Found }
                production:
                  value: { statusCode: 404, message: Not Found }
          x-source: apps/api/src/orders/kiosk-production-availability.guard.ts:18-32 (production), apps/api/src/orders/orders.service.ts:1038-1044 (venue)
        '409':
          $ref: '#/components/responses/OrderConflict'
        '422':
          $ref: '#/components/responses/UnsupportedTaxProfile'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: |
            `Stripe payment is not configured` (no/dummy STRIPE_SECRET_KEY),
            or rate-limit backend down (`Authentication is temporarily unavailable.`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
      x-known-issues:
        - "Historical: the window-display kiosk client was removed on 2026-10-05 (PRD/product-requirements.md section 34.3), and no web kiosk client remains. As found before removal, that client (apps/window-display/src/pages/KioskOrderPage.tsx:112-130, 223-243, and apps/window-display/src/api/menu.ts:27-40) sends NO idempotencyKey, so every submission fails DTO validation with 400. It also computes additive GST (cartTax = round(subtotal*0.15), cartTotal = subtotal + tax, KioskOrderPage.tsx:79-80) and charges that via create-payment-intent, so even with a key the server's GST-inclusive total (== subtotal) would not equal the PaymentIntent amount and step 11 would return 409. It sends no x-csrf-token header and no Bearer token, so CSRF middleware returns 403 before either. Net: the kiosk client cannot currently succeed."
        - "CSRF is bypassed by ANY `Authorization: Bearer <anything>` header; on this public route the token is never verified, so the CSRF check is trivially skippable by non-browser callers."
        - "Idempotent replay does not compare stripePaymentIntentId: a replay with the same key and cart but a different (or unpaid) PaymentIntent returns the original order with 201."
        - "Line pricing (step 4) runs before the replay decision, so a legitimate replay fails with 409/400 if an item was since deleted, made unavailable, or a named modifier removed."
        - "Venue isActive is not checked on order creation (findUnique by id only, orders.service.ts:1038-1044), unlike the menu/table read paths."
        - "Table resolution matches tableId OR tableNumber; if both are supplied and disagree, whichever row Postgres returns first wins (orders.service.ts:1066-1072)."
        - "Kiosk orders are always persisted as serviceMode dine_in, even with no table."
  /kiosk/stripe/connection-token:
    post:
      tags: [kiosk]
      operationId: createStripeTerminalConnectionToken
      summary: Stripe Terminal connection token
      description: |
        Creates a Stripe Terminal connection token with Stripe API version
        `2024-06-20`. No request body is read.

        **Production (Story 12.5):** this route is off whenever `NODE_ENV`
        is anything other than exactly `development` or `test` (including
        unset). A request that gets past CSRF gets a plain `404 Not Found`,
        as for a route that does not exist, before rate limiting, body
        validation and any service, database or Stripe work. There is no
        setting that re-enables it in production.
      x-source: apps/api/src/orders/orders.controller.ts:70-75
      x-implementation: apps/api/src/orders/orders.service.ts:951-962
      x-rate-limit:
        limit: 20
        windowSeconds: 60
        x-source: apps/api/src/orders/orders.controller.ts:71-72
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '201':
          description: Connection token.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [secret]
                properties:
                  secret: { type: string }
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '404':
          description: |
            Production (Story 12.5): the route does not exist. Returned for
            every request that gets past CSRF with a well-formed JSON body
            (or none), empty or invalid, before rate limiting and body
            validation, so never 429 or 503 and never a validation 400. A
            malformed or oversized body is refused (400/413) by the shared
            body parser before routing, as for any path. The body carries no
            route, provider or configuration detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
              example: { statusCode: 404, message: Not Found }
          x-source: apps/api/src/orders/kiosk-production-availability.guard.ts:18-32
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: |
            `Stripe payment is not configured` (STRIPE_SECRET_KEY missing or
            `sk_test_dummy`), `Stripe Terminal is unavailable` (Stripe call
            failed), or rate-limit backend down.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
          x-source: apps/api/src/orders/orders.service.ts:959-986
      x-known-issues:
        - Unauthenticated; anyone who passes CSRF (or sends any Bearer header) can mint Terminal connection tokens for the venue's Stripe account, limited only by the per-IP rate limit.
  /kiosk/stripe/create-payment-intent:
    post:
      tags: [kiosk]
      operationId: createStripePaymentIntent
      summary: Create a card_present NZD PaymentIntent for a client-chosen amount
      description: |
        Calls `stripe.paymentIntents.create({ amount: amountCents, currency:
        'nzd', payment_method_types: ['card_present'] })` (API version
        `2024-06-20`). No capture_method is passed (Stripe default applies).

        **Production (Story 12.5):** this route is off whenever `NODE_ENV`
        is anything other than exactly `development` or `test` (including
        unset). A request that gets past CSRF gets a plain `404 Not Found`,
        as for a route that does not exist, before rate limiting, body
        validation and any service, database or Stripe work. There is no
        setting that re-enables it in production.
      x-source: apps/api/src/orders/orders.controller.ts:77-82
      x-implementation: apps/api/src/orders/orders.service.ts:964-979
      x-rate-limit:
        limit: 20
        windowSeconds: 60
        x-source: apps/api/src/orders/orders.controller.ts:78-79
      security: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentIntentDto'
      responses:
        '201':
          description: PaymentIntent client secret.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [clientSecret]
                properties:
                  clientSecret: { type: ['string', 'null'] }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/CsrfRejected'
        '404':
          description: |
            Production (Story 12.5): the route does not exist. Returned for
            every request that gets past CSRF with a well-formed JSON body
            (or none), empty or invalid, before rate limiting and body
            validation, so never 429 or 503 and never a validation 400. A
            malformed or oversized body is refused (400/413) by the shared
            body parser before routing, as for any path. The body carries no
            route, provider or configuration detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
              example: { statusCode: 404, message: Not Found }
          x-source: apps/api/src/orders/kiosk-production-availability.guard.ts:18-32
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: |
            `Stripe payment is not configured`, `Stripe payment setup is
            unavailable` (Stripe call failed), or rate-limit backend down.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
      x-known-issues:
        - "The amount is chosen by the client and not tied to any cart, venue or order; the server only bounds it to 50..1000000 cents. Correctness relies entirely on POST /kiosk/orders later comparing intent.amount to the server-computed total."
        - "No venue, metadata or idempotency key is attached to the PaymentIntent."
        - "Whether a card_present PaymentIntent created without capture_method reaches status 'succeeded' (required by POST /kiosk/orders) after Terminal processPayment is a Stripe-side behaviour not provable from this repository."
      x-unverified-items:
        - x-unverified: true
          note: Stripe capture behaviour for card_present without explicit capture_method.
  /admin/orders:
    get:
      tags: [admin-orders]
      operationId: listOrders
      summary: List orders (staff, KDS device, elevated tablet)
      description: |
        Returns ALL matching orders, unpaginated, `createdAt DESC`.
        Filter: `venue.organizationId = token.organizationId`, plus
        `venueId` = scoped venue if any, plus when `activeOnly === "true"`
        (exact string) `status IN (confirmed, preparing, ready)`.
      x-source: apps/api/src/orders/orders.controller.ts:54-65
      x-implementation: apps/api/src/orders/orders.service.ts:838-855
      x-roles: { $ref: '#/components/x-staff-order-roles' }
      x-rate-limit: none (RateLimitGuard not applied)
      security:
        - bearerAuth: []
      parameters:
        - name: venueId
          in: query
          required: false
          description: |
            Optional venue filter. NOT validated as a UUID. For device-scoped
            token kinds it must equal the token venue (else 403) and is
            defaulted to it when omitted.
          schema: { type: string }
        - $ref: '#/components/parameters/ActiveOnly'
      responses:
        '200':
          description: Orders with items, table and posSyncRecord.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OrderFull' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
      x-known-issues:
        - No pagination or date bound; returns the organization's entire order history when activeOnly is not "true".
        - activeOnly excludes `pending`, while table occupancy checks treat `pending` as active.
    post:
      tags: [admin-orders]
      operationId: createStaffOrder
      summary: Staff-surface order creation (Order Tablet, elevated tablet, KDS token)
      description: |
        Authenticated, never prepaid. Persisted `status = confirmed`,
        `source = staff`, `confirmedAt = now`,
        `paymentProviderTransactionId = null`, `guests = dto.guests ?? null`.
        Modifiers use the STRICT id-based resolution.

        Controller first calls `resolveVenueScope(user, dto.venueId)`
        (device tokens: 403 on venue mismatch). Then `createStaffOrder`
        (shared with POST /tablet/orders), **exact order**
        (orders.service.ts:256-395):
        1. Venue by id -> 404 `Venue not found`; `venue.organizationId !==
           token org` -> 403 `Venue does not belong to your organization`.
        2. Existing order by `(venueId, idempotencyKey)`.
        3. If none, by `serviceMode`:
           * `dine_in`: `validateTableForOrder(venueId, tableId)`; no tableId
             -> 400 `A valid tableId is required for a dine-in order`;
             unknown/inactive -> 400 `Invalid or inactive table`; occupied ->
             409 `Table <tableNumber> already has an active order`.
           * `takeaway`: `tableId` present -> 400 `A takeaway order must not
             specify a table`. No table is ever assigned.
           * Table 19 validation guard (inert unless
             `TABLE19_LIVE_TEST_ENABLED === 'true'` and NODE_ENV !==
             'production'; then 403s, see x-table19).
        4. Price lines (`resolveOrderItems`, strict) — runs for replays too.
        5. Existing order -> idempotent outcome.
        6. `checkExpectedPrices` -> 409 STALE_PRICE.
        7. `computeTotals` -> 422 on unsupported tax profile.
        8. Persist (same transaction as kiosk; for takeaway a
           `takeawayReference` is minted). Notes prefixed
           `TABLE19-VALIDATION-` only in Table 19 mode.
        9. Audit `CREATE_ORDER` (after: orderId, totalCents, source,
           serviceMode, tableId, takeawayReference), WebSocket broadcast,
           201.
      x-source: apps/api/src/orders/orders.controller.ts:83-90
      x-implementation: apps/api/src/orders/orders.service.ts:256-395
      x-roles: { $ref: '#/components/x-staff-order-roles' }
      x-rate-limit: none (RateLimitGuard not applied)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaffOrderDto'
      responses:
        '201':
          description: New order, or ORIGINAL order on matching replay (same status code).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OrderCreated'
                  - $ref: '#/components/schemas/OrderReplay'
        '400':
          description: |
            DTO validation, or: `A valid tableId is required for a dine-in
            order`; `A takeaway order must not specify a table`; `Invalid or
            inactive table`; `MenuItem with ID <id> not found`; `An order must
            contain at least one item`; any strict modifier error (see
            x-modifier-resolution.strict).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: |
            `Insufficient permissions`; `This device is not authorized for
            the requested venue`; `Venue does not belong to your
            organization`; Table 19 guard messages.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
        '404': { $ref: '#/components/responses/VenueNotFound' }
        '409': { $ref: '#/components/responses/OrderConflict' }
        '422': { $ref: '#/components/responses/UnsupportedTaxProfile' }
      x-known-issues:
        - "`kitchen` is in STAFF_ORDER_ROLES (orders.controller.ts:20-25) and KDS device tokens carry role kitchen (auth.service.ts:138), so any KDS device token can create orders for its venue (and cancel them via PATCH status)."
        - "`owner` role bypasses RolesGuard entirely (roles.guard.ts:41-44)."
        - "Line pricing runs before the replay decision, so a replay can fail if the menu changed (item unavailable/deleted, modifier group/option removed or made unavailable, group constraints changed)."
        - "Venue isActive is not checked on order creation."
        - "Staff tokens (non-device) are not venue-scoped; any venue in the organization is accepted."
  /admin/orders/{id}:
    get:
      tags: [admin-orders]
      operationId: getOrder
      summary: Get one order
      description: |
        `findFirst where id = :id AND venue.organizationId = token org AND
        (venueId = token venue, for device-scoped kinds)`.
        The id is a plain string (not UUID-validated).
      x-source: apps/api/src/orders/orders.controller.ts:67-75
      x-implementation: apps/api/src/orders/orders.service.ts:857-866
      x-roles: { $ref: '#/components/x-staff-order-roles' }
      x-rate-limit: none
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderFull' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          $ref: '#/components/responses/OrderNotFound'
  /admin/orders/{id}/status:
    patch:
      tags: [admin-orders]
      operationId: updateOrderStatus
      summary: Transition order status
      description: |
        1. Load order scoped like GET (org + device venue) -> 404 `Order not found`.
        2. If `newStatus !== oldStatus`, validate against the transition
           table (`components.x-status-transitions`); disallowed -> 409
           `Invalid order status transition from <old> to <new>`.
           **Same-status requests skip validation** and proceed.
        3. If `newStatus === cancelled` and `oldStatus !== cancelled`:
           `LegacyExternalPosHandoff.stopHandoffForCancelledOrder` runs
           BEFORE the status write (see `components.x-cancellation`).
        4. Update `status` and set exactly one timestamp by NEW status:
           confirmed->`confirmedAt`, preparing->`preparingAt`,
           ready->`readyAt`, completed->`completedAt`,
           cancelled->`cancelledAt` (= now). `pending` sets none.
        5. Audit `UPDATE_ORDER_STATUS` with `before: { status }`,
           `after: { status, posDispatchStopped? }` (`posDispatchStopped`
           present only on a real cancellation).
        6. WebSocket `sendOrderUpdate(venueId, updatedOrder)`; return 200.
      x-source: apps/api/src/orders/orders.controller.ts:92-107
      x-implementation: apps/api/src/orders/orders.service.ts:868-949
      x-roles: { $ref: '#/components/x-staff-order-roles' }
      x-rate-limit: none
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
      responses:
        '200':
          description: Updated order (items, table, posSyncRecord).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderFull' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/OrderNotFound' }
        '409':
          description: Disallowed transition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NestError' }
              example:
                statusCode: 409
                message: Invalid order status transition from completed to preparing
                error: Conflict
          x-source: apps/api/src/orders/orders.service.ts:886-893
      x-known-issues:
        - "Same-status updates (e.g. confirmed -> confirmed, cancelled -> cancelled) are accepted: they rewrite the corresponding timestamp, write an audit row and re-broadcast, with no error."
        - "A cancellation whose legacy POS dispatch could not be stopped still succeeds (200); the failure is only visible as posDispatchStopped:false in the audit log and an errorMessage on posSyncRecord."
        - "`kitchen` role (hence every KDS device token) can cancel orders."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT (HS256)
...
  parameters:
    CsrfHeader:
      name: x-csrf-token
      in: header
      required: false
      description: |
        Required unless an `Authorization: Bearer ...` header is present;
        must equal the `csrf_token` cookie (issued by any earlier response).
      schema: { type: string }
      x-source: apps/api/src/common/middleware/csrf.middleware.ts:56-79
...
  schemas:
    NestError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string }
...
    CreateOrderDto:
      type: object
      required: [venueId, items, stripePaymentIntentId, idempotencyKey]
      properties:
        venueId: { type: string, format: uuid }
        tableId: { type: string, format: uuid, description: Optional. }
        tableNumber: { type: string, description: 'Optional. Matched against Table.tableNumber (OR with tableId).' }
        items:
          type: array
          description: '@IsArray, no min size at DTO level; empty -> 400 from service.'
          items: { $ref: '#/components/schemas/CreateOrderItemDto' }
        notes: { type: string }
        stripePaymentIntentId: { type: string, minLength: 1 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 255 }
      x-source: apps/api/src/orders/dto/create-order.dto.ts:100-142
...
    OrderCreated:
      description: |
        Response of a NEW order from any create route:
        `include: { items: true, posSyncRecord: true }` — NO `table`.
      x-source: apps/api/src/orders/orders.service.ts:1850-1866
      allOf:
        - $ref: '#/components/schemas/OrderBase'
        - type: object
          required: [items, posSyncRecord]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OrderItem' }
            posSyncRecord:
              x-legacy: true
              oneOf:
                - $ref: '#/components/schemas/POSSyncRecord'
                - type: 'null'
...